Search guide reformat and pretty-print
Indent and reformat the guide using Emacs HTML mode indent-for-tab-command
and fill-paragraph commands.  There has been absolutely no content change in
this patch.  There are a few mismatch tags and a couple of unescaped
greater-than characters.  They will be fixed in the next patch, because
otherwise it is too difficult to isolate the changes.