Search guide webdoc reformat and pretty-print (3/3)
Move paragraphs around so the two lists of the section ``How to search for
publications by a given author'' become a single list again.  I introduced
this error during my previous commit.  The reason was that this (quite long)
section was split in two between the original English and French versions,
so in the middle of the English text, without closing the <ol>, started the
French version, and after this translation, the English text continued,
followed by the French translation.  and I wrongly understood that the
missing </ol> was between the first two texts.  I've joined all English text
together, in a single <ol> list, and done the same with the other existing
translations.